Add crossOrigin member to CollectedClientData.
This member indicates that an operation was performed in a cross-origin
context. This embodies three decisions:

1) That the top-level origin should not be exposed. I don't believe that
there are other APIs which allow an iframe to discover the top-level
origin and adding one here thus seems unwise. We expect that the
authenticated source of messages using postMessage will inform an iframe
of the requesting origin.

2) That we aren't going to break the existing CollectedClientData in a
cross-origin context. I.e. an RP which doesn't know about this addition
will continue to consider the CollectedClientData to be valid. We are
doing this so as not to break existing server-side WebAuthn libraries.

3) We are not adding a pre-request guard. I.e. there's no
|crossOriginOk| boolean when making a request. Other Web APIs don't do
this and there are other mechanisms for a page to refuse to load in an
iframe.